Fix a small import issue
Summary:
Hi Charles,

Apologies for the confusing write-up. So, I've found a small import issue with the client. It manifests itself only when we're importing the client and the main inbox package is in the path. Because of the way pkgutil.extend_path works, the client's __init__.py is not executed and therefore APIClient is not available.

To fix this I added an APIClient import into inbox.client's __init__.py and I changed the main inbox package's __init__.py to this:

```
from pkgutil import extend_path
__path__ = extend_path(__path__, __name__)

try:
    from inbox.client import APIClient
    __all__ = ['APIClient']
except ImportError:
    pass
```

So from inbox import APIClient works for everyone now.

Test Plan: Tested imports.

Reviewers: charles

Reviewed By: charles

Differential Revision: https://review.inboxapp.com/D399