[Python-SDK] Add support for querying account status
Summary: This diff is pretty straightforward, with the exception of a little bit of metaprogramming in restful_model_collections.py. The problem is that the original code assumed we would only query resources under `/n`. To solve this I added a new field to every resource class, `api_namespace` to store whether the resource is under `/a` or `/n`.

Test Plan: Tested manually.

Reviewers: kav-ya, spang, emfree

Reviewed By: emfree

Projects: #python-sdk

Differential Revision: https://review.inboxapp.com/D1018