[62077] Fix adding a tracking object to an existing `draft` (#153)
The Python SDK creates a special payload specifically for sending an already-existing `draft` object that extracts the draft ID and the draft's version and sends that as a payload. However, it does not check if a tracking object exists, and thus if a user were to save a draft, and come add a tracking object to it (or if the user saves the draft with the tracking object first) and tries to send it after the save, the SDK never sends the tracking object in the payload. This PR adds a check to add a tracking object if present in the `draft`.