[62132] Fix issue when converting offset-aware `datetime` objects to `timestamp` (#154)
Wherever we accept a date as a filtering parameter, we usually convert it from a `datetime` object to a `timestamp` object. However the way we convert it does not take into consideration if the `datetime` is offset-aware or not (timezone), and thus would throw an error when it's attempted. Now we check if the `datetime` is context-aware, and if it is, we convert it properly to being an offset-naive object before converting it to a timestamp.