[62378] Updated credential naming to be consistent with the API and other SDKs (#159)
In the other Nylas components like the API and Dashboard, as well as the other SDKs, the credentials are referred to as client_id and client_secret however this SDK was referring to these credentials as app_id and app_secret. This causes some unnecessary confusion, and this change addresses this by switching all references of app_id and app_secret to client_id and client_secret.