Python middleware infrastructure for use cases (#232)
This PR lays the foundation for the Python SDK to build middleware to integrate into your backend application with routes to some endpoints that would help with integrating authentication and webhook event. Currently provided routes:

    /nylas/generate-auth-url - Endpoint that takes an optional email address and success redirect path, returns a URL used for authentication
    /nylas/exchange-mailbox-token - Endpoint that takes a token from the JSON and exchanges that code for token, then calls a user-defined callback

This PR also includes support for local webhook development. The SDK will create a tunnel connection to a websocket server and registers it as a webhook callback to your Nylas account.