[v6] Fix hasattr logic in RestfulModel (#245)
⚠️ This is a breaking change ⚠️

This PR fixes the broken hasattr logic in the RestfulModel superclass. RestfulModel extends dict and sets __getattr__ to dict.get() which never throws if a key is not found, it returns None as a default. This breaks the underlying hasattr function, because the way hasattr works is it calls getattr and if it throws a ArgumentError then it returns false (as described in #220). However, since dict.get() never errors out (returns None if not found), hasattr always returns true.

Now, we set __getattr__ to dict[key] and throws an error if the key does not exist. This will have a cascading effect as it will no longer be safe to just access a variable like object.attr without checking hasattr(object, 'attr') first. This PR will resolve #220 and be released in the next major release as it is a breaking change